+++ test/step-definitions/template.step.ts
@@ -1,9 +1,9 @@
 import path from 'path'
 import { loadFeature, defineFeature } from 'jest-cucumber'
 import { Server, ServerInjectResponse } from '@hapi/hapi'
-import Config from '../../src/shared/config'
+import Config from '~/shared/config'
 
-import ThirdPartyAPIAdapterService from '../../src/server'
+import ThirdPartyAPIAdapterService from '~/server'
 
 const featurePath = path.join(__dirname, '../features/template.scenario.feature')
 const feature = loadFeature(featurePath)
@@ -43,27 +43,6 @@ defineFeature(feature, (test): void => {
       expect(response.statusCode).toBe(200)
       expect(healthResponse.status).toEqual('OK')
       expect(healthResponse.uptime).toBeGreaterThan(1.0)
-    })
-  })
-
-  test('Hello', ({ given, when, then }): void => {
-    given('thirdparty-api-adapter server', async (): Promise<Server> => {
-      server = await ThirdPartyAPIAdapterService.run(Config)
-      return server
-    })
-
-    when('I get \'Hello\' response', async (): Promise<ServerInjectResponse> => {
-      const request = {
-        method: 'GET',
-        url: '/hello'
-      }
-      response = await server.inject(request)
-      return response
-    })
-
-    then('I see \'Hello world\'', (): void => {
-      expect(response.statusCode).toBe(200)
-      expect(response.result).toEqual({ hello: 'world' })
     })
   })
 })