+++ CHANGELOG.md
@@ -1,4 +1,18 @@
 # Changelog: [mojaloop/thirdparty-api-adapter](https://github.com/mojaloop/thirdparty-api-adapter)
+## [11.0.0](https://github.com/mojaloop/thirdparty-api-adapter/compare/v10.4.2...v11.0.0) (2020-08-02)
+
+
+### Features
+
+* POST thirdpartyRequests/transactions ([db6738c](https://github.com/mojaloop/thirdparty-api-adapter/commit/db6738c3cc11319e804a069d1651188e2dd32c7f))
+
+## [11.0.0](https://github.com/mojaloop/thirdparty-api-adapter/compare/v10.4.2...v11.0.0) (2020-08-02)
+
+
+### Features
+
+* POST thirdpartyRequests/transactions ([db6738c](https://github.com/mojaloop/thirdparty-api-adapter/commit/db6738c3cc11319e804a069d1651188e2dd32c7f))
+
 ### 10.4.1 (2020-06-22)